Fix another Doom bug, this one of my own making
finalecount must be an int. Previously it was changed to size_t
to avoid a compiler warning, but that breaks the finale behavior.
When the finale text is displayed, the calculation that tries
to figure out how much text to draw in F_TextWrite() goes bananas
and draws the whole text in the first frame, which is visible right
after the screen wipe. It also breaks the bunny scroll at the end
of the last map in the registered version of the game.

+++ software/firmware/linuxdoom-1.10/f_finale.c
@@ -53,7 +53,7 @@ __attribute__((section(".ram7")))
 int		finalestage;
 
 __attribute__((section(".ram7")))
-size_t		finalecount;
+int		finalecount;
 
 #define	TEXTSPEED	3
 #define	TEXTWAIT	250
@@ -242,7 +242,8 @@ void F_Ticker (void)
     if ( gamemode == commercial)
 	return;
 		
-    if (!finalestage && finalecount>strlen (finaletext)*TEXTSPEED + TEXTWAIT)
+    if (!finalestage &&
+        (size_t)finalecount>strlen (finaletext)*TEXTSPEED + TEXTWAIT)
     {
 	finalecount = 0;
 	finalestage = 1;
